requirements.txt
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
pyarrow>=12.0.0
pytz>=2023.0
